Add support for versioning in dependency name
Pick the dependency version by wildcard, then filter it by #dep.vsn_regex
The highest number will be picked or with fallback to app name (`<depsdir>/#dep.app`).

Motivation for this are cases where rebar does not have sole responsibility
for managing and building dependencies and a the stricter <app>-<vsn>
format is used for dependencies.